unsigned -> NSUInteger
git-svn-id: svn://qnap.local/TwoTerm/branches/frameless@3097 5590a31f-7b70-45f8-8c82-aa3a8e5f4507

+++ NewTerminalWindowController.m
@@ -133,9 +133,9 @@ enum {
 {
     
     NSMenuItem *item = [_terminalTypeButton selectedItem];
-    unsigned tag = [item tag];
+    NSUInteger tag = [item tag];
     
-    Class klass = [EmulatorManager emulatorForTag: tag];
+    Class klass = [EmulatorManager emulatorForTag: (unsigned)tag];
     
     if (klass)
     {
